Issue-8: Moved FocusResizeActivity to separate module.
