+++ app/build.gradle
@@ -31,8 +31,7 @@ dependencies {
 
     implementation 'androidx.appcompat:appcompat:1.1.0'
 
-    // FocusResize RecyclerView
-    implementation 'com.borjabravo:focusresize:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
 
     // VegaLayoutManager
     implementation 'com.stone.vega.library:VegaLayoutManager:1.0.6'
@@ -44,6 +43,7 @@ dependencies {
     implementation project(path: ':swipeback')
     implementation project(path: ':scalinglayout')
     implementation project(path: ':expandablelayout')
+    implementation project(path: ':focusresize')
 }
 
 // For resolve library conflict(dou navigation drawer)
