+++ app/src/main/AndroidManifest.xml
@@ -15,11 +15,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <activity
-            android:name=".recycler.focusresize.FocusResizeActivity"
-            android:label="@string/focus.resize" />
-
         <activity
             android:name=".recycler.vega.VegaManagerActivity"
             android:label="@string/vega.layout.manager" />
