+++ app/src/main/res/layout/activity_main.xml
@@ -46,7 +46,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="focusResize"
-            android:text="@string/focus.resize" />
+            android:text="@string/focus_resize_example" />
 
         <Button
             android:layout_width="match_parent"
