+++ focusresize/src/main/res/layout/a_focus_resize.xml
