+++ focusresize/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="focus_resize_title">FocusResize</string>
+    <string name="focus_resize_example">FocusResize example</string>
+</resources>
