+++ scalinglayout/src/main/res/layout/a_scaling_collapsing.xml
@@ -22,7 +22,6 @@
                 android:layout_width="match_parent"
                 android:layout_height="200dp"
                 android:scaleType="centerCrop"
-                android:src="@drawable/focus_image_1"
                 app:layout_collapseMode="parallax" />
 
             <androidx.appcompat.widget.Toolbar
